refactor(classes): group classes into subfolders like Ganyu
Split the flat src/structures/classes into concern subfolders mirroring Ganyu:
client/ (Stelle, WebhookClient), components/ (EmbedPaginator), database/
(Database, PrismaService, Controller), modules/ (Cache), manager/ (Manager,
Store).

The #stelle/classes/* alias (package.json + tsconfig) is a wildcard, so no
config change is needed — only the import specifiers move: cross-folder imports
use the alias with its subfolder, same-folder ones stay relative.
