+++ src/module.ts
@@ -1,6 +1,6 @@
 import type { LyricsResult } from "hoshimi";
 import type { ParseClient, ParseGlobalMiddlewares, ParseLocales } from "seyfert";
-import type { Stelle } from "#stelle/classes/Stelle.js";
+import type { Stelle } from "#stelle/classes/client/Stelle.js";
 import type { Options, TrackUser } from "#stelle/types";
 import type { StelleContext } from "#stelle/utils/functions/internal/context.js";
 import type English from "./locales/en-US.js";
