+++ src/structures/classes/database/Database.ts
@@ -1,11 +1,11 @@
 import type { UsingClient } from "seyfert";
+import { Cache } from "#stelle/classes/modules/Cache.js";
 import { LocaleController } from "#stelle/controllers/locale.js";
 import { PlayerController } from "#stelle/controllers/player.js";
 import { PlaylistController } from "#stelle/controllers/playlist.js";
 import { PrefixController } from "#stelle/controllers/prefix.js";
 import { RequestsController } from "#stelle/controllers/requests.js";
 import type { PrismaClient } from "#stelle/prisma";
-import { Cache } from "./Cache.js";
 import { PrismaService } from "./PrismaService.js";
 
 /**
