+++ src/structures/classes/manager/Manager.ts
