+++ src/structures/classes/manager/Store.ts
