+++ src/structures/classes/modules/Cache.ts
