+++ src/structures/controllers/requests.ts
@@ -1,4 +1,4 @@
-import { Controller } from "#stelle/classes/Controller.js";
+import { Controller } from "#stelle/classes/database/Controller.js";
 import type { guildRequestChannel } from "#stelle/prisma";
 
 /**
