+++ src/structures/utils/functions/components/playlist.ts
@@ -19,7 +19,7 @@ import { EmbedColors, type PermissionStrings } from "seyfert/lib/common/index.js
 import type { CreateComponentCollectorResult } from "seyfert/lib/components/handler.js";
 import type { PermissionsBitField } from "seyfert/lib/structures/extra/Permissions.js";
 import { ButtonStyle, MessageFlags, TextInputStyle } from "seyfert/lib/types/index.js";
-import { EmbedPaginator } from "#stelle/classes/EmbedPaginator.js";
+import { EmbedPaginator } from "#stelle/classes/components/EmbedPaginator.js";
 import type { userPlaylist } from "#stelle/prisma";
 import type { PermissionNames } from "#stelle/types";
 import { ManageButtonIdentifiers, SaveButtonCustomIds, SaveButtonIdentifiers, type TrackUser } from "#stelle/types";
